fix: address hygiene findings in CalDAV integration
Split oversized source files:
- tools.ts (812 lines) -> tool_definitions.ts, tool_query_handlers.ts,
  tool_mutation_handlers.ts, tool_reports.ts, tool_handlers.ts (barrel), tools.ts (dispatch)
- ical.ts (521 lines) -> ical_parse.ts, ical_generate.ts, ical_freebusy.ts, ical.ts (barrel)
- client.ts (374 lines) -> client.ts (trimmed), client_xml.ts (XML parsing + helpers)

Split oversized test files:
- ical_test.ts (450 lines) -> ical_parse_test.ts, ical_generate_test.ts
- discovery_test.ts (364 lines) -> discovery_endpoint_test.ts, discovery_listing_test.ts
- client_test.ts (346 lines) -> client_read_test.ts, client_write_test.ts

Other fixes:
- Add @module JSDoc tags to all CalDAV test files
- Add caldav_executor.ts to src/gateway/startup/factory/mod.ts
- Fix pre-existing test bugs: unfoldLines test input, Response(204) null body
