+++ src/gateway/startup/factory/mod.ts
@@ -10,3 +10,4 @@ export { buildGoogleExecutor } from "./google_executor.ts";
 export type { WebToolsResult } from "./web_tools.ts";
 export { buildWebTools } from "./web_tools.ts";
 export { buildSchedulerConfig } from "./scheduler_config.ts";
+export { buildCalDavExecutor } from "./caldav_executor.ts";
