+++ src/integrations/caldav/client.ts
@@ -10,12 +10,17 @@
 import type {
   CalDavClientInterface,
   CalDavClientResult,
-  PropfindResource,
   PropfindResponse,
   PutResponse,
-  ReportResource,
   ReportResponse,
 } from "./types.ts";
+import {
+  parsePropfindXml,
+  parseReportXml,
+  isMultiStatus,
+  buildErrorResult,
+  errorMessage,
+} from "./client_xml.ts";
 import { createLogger } from "../../core/logger/logger.ts";
 
 const log = createLogger("caldav:client");
@@ -119,58 +124,6 @@ async function executePropfind(
   }
 }
 
-/** Parse PROPFIND multistatus XML response. */
-function parsePropfindXml(xml: string): PropfindResource[] {
-  const resources: PropfindResource[] = [];
-  const responseRegex = /<(?:d:|D:|DAV:)?response>([\s\S]*?)<\/(?:d:|D:|DAV:)?response>/gi;
-  let responseMatch: RegExpExecArray | null;
-
-  while ((responseMatch = responseRegex.exec(xml)) !== null) {
-    const responseBlock = responseMatch[1];
-    const href = extractTagValue(responseBlock, "href") ?? "";
-    const properties: Record<string, string> = {};
-
-    // Extract common properties
-    const displayName = extractTagValue(responseBlock, "displayname");
-    if (displayName) properties["displayname"] = displayName;
-
-    const calendarColor = extractTagValue(responseBlock, "calendar-color");
-    if (calendarColor) properties["calendar-color"] = calendarColor;
-
-    const calendarDescription = extractTagValue(responseBlock, "calendar-description");
-    if (calendarDescription) properties["calendar-description"] = calendarDescription;
-
-    const getctag = extractTagValue(responseBlock, "getctag");
-    if (getctag) properties["getctag"] = getctag;
-
-    const getetag = extractTagValue(responseBlock, "getetag");
-    if (getetag) properties["getetag"] = getetag;
-
-    const currentUserPrincipal = extractHrefFromTag(responseBlock, "current-user-principal");
-    if (currentUserPrincipal) properties["current-user-principal"] = currentUserPrincipal;
-
-    const calendarHomeSet = extractHrefFromTag(responseBlock, "calendar-home-set");
-    if (calendarHomeSet) properties["calendar-home-set"] = calendarHomeSet;
-
-    const resourcetype = responseBlock.includes("calendar")
-      ? (responseBlock.includes("<c:calendar") || responseBlock.includes("<cal:calendar") || responseBlock.match(/<[^>]*calendar[^>]*\/>/i))
-        ? "calendar"
-        : ""
-      : "";
-    if (resourcetype) properties["resourcetype"] = resourcetype;
-
-    const status = extractTagValue(responseBlock, "status");
-
-    resources.push({
-      href,
-      properties,
-      ...(status ? { status } : {}),
-    });
-  }
-
-  return resources;
-}
-
 // ─── REPORT ───────────────────────────────────────────────────────────────────
 
 /** Execute a REPORT request. */
@@ -207,26 +160,6 @@ async function executeReport(
       error: { status: 0, message: `REPORT network error: ${errorMessage(err)}` },
     };
   }
-}
-
-/** Parse REPORT multistatus XML response. */
-function parseReportXml(xml: string): ReportResource[] {
-  const resources: ReportResource[] = [];
-  const responseRegex = /<(?:d:|D:|DAV:)?response>([\s\S]*?)<\/(?:d:|D:|DAV:)?response>/gi;
-  let responseMatch: RegExpExecArray | null;
-
-  while ((responseMatch = responseRegex.exec(xml)) !== null) {
-    const block = responseMatch[1];
-    const href = extractTagValue(block, "href") ?? "";
-    const etag = (extractTagValue(block, "getetag") ?? "").replace(/"/g, "");
-    const calendarData = extractTagValue(block, "calendar-data") ?? "";
-
-    if (calendarData) {
-      resources.push({ href, etag, calendarData });
-    }
-  }
-
-  return resources;
 }
 
 // ─── PUT ──────────────────────────────────────────────────────────────────────
@@ -314,61 +247,4 @@ async function executeDelete(
       error: { status: 0, message: `DELETE network error: ${errorMessage(err)}` },
     };
   }
-}
-
-// ─── XML Helpers ──────────────────────────────────────────────────────────────
-
-/**
- * Extract the text content of an XML tag, ignoring namespace prefixes.
- *
- * Handles `<d:tag>value</d:tag>`, `<D:tag>value</D:tag>`, and `<tag>value</tag>`.
- */
-function extractTagValue(xml: string, tagName: string): string | null {
-  const regex = new RegExp(
-    `<(?:[a-zA-Z][a-zA-Z0-9]*:)?${tagName}[^>]*>([\\s\\S]*?)<\/(?:[a-zA-Z][a-zA-Z0-9]*:)?${tagName}>`,
-    "i",
-  );
-  const match = regex.exec(xml);
-  return match ? match[1].trim() : null;
-}
-
-/** Extract an href from within a nested tag (e.g., `<d:current-user-principal><d:href>...</d:href></d:current-user-principal>`). */
-function extractHrefFromTag(xml: string, tagName: string): string | null {
-  const outerRegex = new RegExp(
-    `<(?:[a-zA-Z][a-zA-Z0-9]*:)?${tagName}[^>]*>([\\s\\S]*?)<\/(?:[a-zA-Z][a-zA-Z0-9]*:)?${tagName}>`,
-    "i",
-  );
-  const outerMatch = outerRegex.exec(xml);
-  if (!outerMatch) return null;
-  return extractTagValue(outerMatch[1], "href");
-}
-
-// ─── Shared Utilities ─────────────────────────────────────────────────────────
-
-/** Check if an HTTP status is 207 Multi-Status. */
-function isMultiStatus(status: number): boolean {
-  return status === 207;
-}
-
-/** Build an error result from an HTTP response. */
-async function buildErrorResult<T>(
-  response: Response,
-  method: string,
-): Promise<CalDavClientResult<T>> {
-  const text = await response.text().catch(() => "");
-  const message = `${method} failed with status ${response.status}: ${text.substring(0, 200)}`;
-  log.warn(`CalDAV ${method} error`, {
-    operation: method.toLowerCase(),
-    status: response.status,
-    err: { message },
-  });
-  return {
-    ok: false,
-    error: { status: response.status, message },
-  };
-}
-
-/** Extract error message from unknown error. */
-function errorMessage(err: unknown): string {
-  return err instanceof Error ? err.message : String(err);
 }
