+++ src/integrations/caldav/tool_handlers.ts
@@ -1,0 +1,21 @@
+/**
+ * CalDAV tool handler barrel.
+ *
+ * Re-exports all tool handler functions from query and mutation modules.
+ *
+ * @module
+ */
+
+export {
+  resolveCalendarUrl,
+  executeCalendarsList,
+  executeEventsList,
+  executeEventsGet,
+  executeFreeBusy,
+} from "./tool_query_handlers.ts";
+
+export {
+  executeEventsCreate,
+  executeEventsUpdate,
+  executeEventsDelete,
+} from "./tool_mutation_handlers.ts";
