+++ tests/integrations/caldav/auth_test.ts
@@ -1,4 +1,6 @@
 /**
+ * @module
+ *
  * CalDAV authentication tests.
  *
  * Tests credential resolution from keychain and Authorization header building.
